% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{author_data}
\alias{author_data}
\title{Author Data for Biblical Books}
\format{
A data frame with 66 rows and 6 variables:
\describe{
\item{author}{Character. Name of the author (e.g., "Moses").}
\item{books}{Character. Name of the book(s) authored.}
\item{section}{Character. Section of the Bible (e.g., "Law", "History").}
\item{date}{Character. Approximate date of authorship (e.g., "c. 1445-1405 BC").}
\item{testament}{Character. Testament classification ("Old Testament" or "New Testament").}
\item{language}{Character. Language in which the book was originally written (e.g., "Hebrew").}
}
}
\usage{
author_data
}
\description{
A dataset containing information about the authors, books, sections, and languages of the Bible.
}
\details{
This dataset summarizes metadata about biblical authors, sections, and book classifications.
}
\keyword{datasets}
