% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by.R
\name{by_author}
\alias{by_author}
\title{Retrieve standardized book names based on biblical authors}
\usage{
by_author(author, testament)
}
\arguments{
\item{author}{a character vector or length 1 specifying the biblical author in the Old or New Testament}

\item{testament}{a character vector of length 1 specifying the testament/time period in which the author is found}
}
\value{
a character vector of standardized book names to be passed to other functions
}
\description{
Collect book names from RASB Bible, Leningrad Codex, or Septuagint based on biblical authors
outlined in 'author_data' dataset.
\itemize{
\item Moses: Genesis – Deuteronomy
\item Joshua: Joshua
\item Unkown: Judges, Ruth, 1 & 2 Kings, 1 & Chronicles, Esther and Job
\item Samuel/Unknown: 1 & 2 Samuel
\item Ezra: Ezra
\item Nehemiah: Nehemiah
\item David/Various: Psalms
\item Solomon: Proverbs, Ecclesiastes, Song of Solomon
\item Isaiah: Isaiah
\item Jeremiah: Jeremiah and Lamentations
\item Ezekiel: Ezekiel
\item Daniel: Daniel
\item Hosea: Hosea
\item Joel: Joel
\item Amos: Amos
\item Obadiah: Obadiah
\item Jonah: Jonah
\item Micah: Micah
\item Nahum: Nahum
\item Habakkuk: Habakkuk
\item Zephaniah: Zephaniah
\item Haggai: Haggai
\item Zechariah: Zechariah
\item Malachi: Malachi
\item Matthew: Gospel of Matthew
\item Mark: Gospel of Mark
\item Luke: Gospel of Luke and Acts
\item John: Gospel of John, Johanine epistles and Revelation
\item Paul: Pauline epistles
\item James: James
\item Peter: 1 and 2 Peter
\item Jude: Jude
\item Unknown: Hebrews}
}
\keyword{internal}
