% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by.R
\name{by_date}
\alias{by_date}
\title{Retrieve standardized book names based on probable writing period}
\usage{
by_date(date_range, testament)
}
\arguments{
\item{date_range}{a character vector or length 1, written as a date range, "1445 - 1405 BC", specifying the probable writing period of the Old or New Testament books}

\item{testament}{a character vector of length 1 specifying the testament corresponding to the provided date range}
}
\value{
a character vector of standardized book names to be passed to other functions
}
\description{
Collect book names from RASB Bible, Leningrad Codex, or Septuagint based on the dates
outlined in 'author_data' dataset.
}
\keyword{internal}
