% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by.R
\name{by_section}
\alias{by_section}
\title{Retrieve standardized book names based on biblical narrative styles or genres}
\usage{
by_section(section, testament)
}
\arguments{
\item{section}{a character vector or length 1 specifying the narrative section or genre in the Old or New Testament}

\item{testament}{a character vector of length 1 specifying the testament in which the section is found}
}
\value{
a character vector of standardized book names to be passed to other functions
}
\description{
Collect book names from RASB Bible, Leningrad Codex, or Septuagint based on narrative styles
outlined in 'author_data' dataset.
\itemize{
\item Law: Genesis – Deuteronomy
\item History: Joshua – Esther
\item Wisdom: Job – Songs of Solomon
\item Major Prophets: Isaiah - Daniel
\item Minor Prophets: Hosea – Malachi
\item Gospels: Matthew – John
\item Pauline Epistles: Romans – Philemon
\item General Epistles: Hebrews – Jude
\item Apocalyptic: Revelation}
}
\examples{
by_section("Gospels","new")

}
