% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_size.R
\name{get_folder_size}
\alias{get_folder_size}
\title{Calculate the Size of a Folder}
\usage{
get_folder_size(
  folder,
  units = "Kb",
  recursive = TRUE,
  omit_folders = TRUE,
  ...
)
}
\arguments{
\item{folder}{A string specifying the name of the folder whose size is to be calculated.}

\item{units}{A string specifying the units for the size calculation. Options are:
\code{"B"} (bytes), \code{"Kb"} (kilobytes), \code{"Mb"} (megabytes), or \code{"Gb"} (gigabytes). Defaults to \code{"Kb"}.}

\item{recursive}{A logical value indicating whether to include subdirectories when listing files. Defaults to \code{TRUE}.}

\item{omit_folders}{A logical value indicating whether to exclude specific folders (e.g., \code{.git}, \code{.quarto}, \code{renv}) from the search. Defaults to \code{TRUE}.}

\item{...}{further arguments to be passed to \code{is_file} or \code{omit_folders}}
}
\value{
A data frame containing:
\itemize{
\item \code{folder}: The name of the folder.
\item \code{size}: The total size of the folder in the specified units.
\item \code{files}: The total number of files in the folder.
\item \code{extensions}: A list of unique file extensions in the folder.
}
}
\description{
Internal function to calculate the total size of a folder, the number of files,
and the unique file extensions within the folder.
}
\details{
The function allows the user to calculate folder sizes while excluding specific folders
such as \code{renv} or \code{.git} to optimize performance. If multiple folders with the same name
are found, the user is prompted to select one.
}
\keyword{internal}
