% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_size.R
\name{is_file}
\alias{is_file}
\title{Check if a Name is a File or a Folder}
\usage{
is_file(name, dir, ignore_dot_folders = TRUE, ...)
}
\arguments{
\item{name}{A string specifying the name to check.}

\item{dir}{A string specifying the directory to search within.}

\item{ignore_dot_folders}{A logical value indicating whether to ignore
hidden folders (e.g., \code{.git}, \code{.quarto}) during the check. Defaults to \code{TRUE}.}
}
\value{
A logical value: \code{TRUE} if the name is a file, \code{FALSE} otherwise.
}
\description{
Internal function to verify whether a given name corresponds to a file
or a folder in the specified directory.
}
\details{
This function searches for the specified name in the directory and distinguishes
between files and folders. If the name corresponds to a folder, a warning is issued.
}
\keyword{internal}
