% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logos-package.R
\docType{package}
\name{logos-package}
\alias{logos}
\alias{logos-package}
\title{logos: Access to the Leningrad Codex Version of the Old Testament,the Greek New Testament, and the Revised American Standard Bible}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Access to the Greek New Testament (27 books) and the Old Testament (39 books) and allow users to do textual analysis on the data. The New and Old Testament have been provided in their original languages, Greek and Hebrew, respectively. Additionally, the Revised American Standard Bible is also provided for users who'd rather use a word–for–word modern English translation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jpmonteagudo28/logos}
  \item Report bugs at \url{https://github.com/jpmonteagudo28/logos/issues}
}

}
\author{
\strong{Maintainer}: JP Monteagudo \email{jpmonteagudo2014@gmail.com} (\href{https://orcid.org/0009-0003-6465-6658}{ORCID}) [copyright holder]

}
\keyword{internal}
