% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{new_testament}
\alias{new_testament}
\title{The Society of Biblical Literature Greek New Testament}
\format{
A data frame with 7,939 rows and 5 variables:
\describe{
\item{book}{Character. Name of the book (e.g., "1Cor").}
\item{chapter}{Character. Chapter number.}
\item{verse}{Character. Verse number.}
\item{text}{Character. Text of the verse in Greek.}
\item{greek_name}{Character. Greek name of the book.}
}
}
\usage{
new_testament
}
\description{
A dataset containing the text of the New Testament in Greek with book metadata.
The Society of Biblical Literature, in keeping with its mission to foster
biblical scholarship, is pleased to sponsor, in association with Logos Bible
Software, a new, critically edited edition of the Greek New Testament.
}
\details{
This dataset provides the Greek text of the New Testament along with book and chapter metadata.
}
\keyword{datasets}
