% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{old_testament}
\alias{old_testament}
\title{Old Testament Dataset}
\format{
A data frame with 23,213 rows and 5 variables:
\describe{
\item{book}{Character. Name of the book (e.g., "Amos").}
\item{chapter}{Numeric. Chapter number.}
\item{verse}{Numeric. Verse number.}
\item{text}{Character. Text of the verse in Hebrew.}
\item{hebrew_names}{Character. Hebrew name of the book.}
}
}
\usage{
old_testament
}
\description{
A dataset containing the text of the Old Testament in Hebrew with book metadata.
The Unicode/XML Leningrad Codex (UXLC) is a transcription of the Leningrad Codex (LC) into a
modern computer format (Unicode, XML). The UXLC text is a fork of the Groves Center's Westminster
Leningrad Codex \href{https://www.bible.com/versions/904-wlc-westminster-leningrad-codex-groves-center-version}{ WLC 4.20, 2016}
}
\details{
This dataset provides the Hebrew text of the Old Testament along with book and chapter metadata.
}
\keyword{datasets}
