% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rasb_bible}
\alias{rasb_bible}
\title{RASB Bible}
\format{
A data frame with 31,102 rows and 4 variables:
\describe{
\item{book}{Character. Name of the book (e.g., "Gen").}
\item{chapter}{Integer. Chapter number.}
\item{verse}{Integer. Verse number.}
\item{text}{Character. Text of the verse.}
}
}
\usage{
rasb_bible
}
\description{
A public domain dataset containing the text of the Bible from the Revised American Standard Bible (RASB).
}
\details{
This dataset provides the full English text of the Bible from the RASB translation.
}
\keyword{datasets}
