% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Select columns of a data frame}
\usage{
select(.data, ...)
}
\arguments{
\item{.data}{A \code{data.frame}.}

\item{...}{The name(s) of the column(s) to select.}
}
\description{
Select columns of a data frame
}
\examples{
select(mtcars, mpg, cyl)
select(mtcars, MilesPerGallon = mpg, Cylinders = cyl)
mtcars |> select(mpg)
mtcars |> select(mpg, cyl)

}
