% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{standardize_testament}
\alias{standardize_testament}
\title{Standardize Bible testaments to a common format}
\usage{
standardize_testament(testament)
}
\arguments{
\item{testament}{A character vector of testament names.}
}
\value{
A character vector of standardized testament names.
}
\description{
This function converts testament to their standardized abbreviations using a predefined lookup table.
}
