% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_book}
\alias{validate_book}
\title{Validate book names against a reference list}
\usage{
validate_book(book, book_list)
}
\arguments{
\item{book}{A character vector, list, or data frame containing book names.}

\item{book_list}{A character vector of valid book names.}
}
\value{
A character vector of validated or suggested book names.
}
\description{
This function checks if the provided book names exist in a given list and suggests the closest match if not.
}
\keyword{internal}
